# lib/api-zod/src/generated/types/catalogStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Data Catalog API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CatalogStats {
  total_datasets: number;
  total_tables: number;
  locations_count: number;
  project_id: string;
  /** @nullable */
  last_updated?: string | null;
}