# lib/api-zod/src/generated/types/dataset.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Data Catalog API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Dataset {
  dataset_id: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  description_he?: string | null;
  location: string;
  tables_count: number;
  created_at: string;
}