# lib/api-zod/src/generated/types/datasetList.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Data Catalog API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Dataset } from './dataset';

export interface DatasetList {
  datasets: Dataset[];
  total: number;
}